BAH-4858 | Fix. reuse observation mocks to resolve sonar issue for code duplication
