+++ packages/bahmni-widgets/src/observations/__mocks__/observationTestData.ts
@@ -816,3 +816,133 @@ export const mockObservationWithoutObservationValue = {
   display: 'Notes Only',
   observationValue: undefined,
 };
+
+export const mockObservationWithFormPath: Observation = {
+  resourceType: 'Observation',
+  id: 'obs-1',
+  status: 'final',
+  code: {
+    coding: [
+      {
+        code: '5087AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
+        display: 'Pulse',
+      },
+    ],
+    text: 'Pulse',
+  },
+  subject: {
+    reference: 'Patient/patient-uuid',
+  },
+  encounter: {
+    reference: 'Encounter/encounter-1',
+  },
+  effectiveDateTime: '2026-07-20T09:59:41+00:00',
+  valueQuantity: {
+    value: 76,
+    unit: 'beats/min',
+  },
+  extension: [
+    {
+      url: 'http://fhir.bahmni.org/ext/observation/form-namespace-path',
+      valueString: 'Bahmni^Vitals (6 years or older).1/17-0',
+    },
+  ],
+};
+
+export const mockObservationsForVitals: Observation[] = [
+  mockObservationWithFormPath,
+  {
+    ...mockObservationWithFormPath,
+    id: 'obs-vitals-2',
+    code: {
+      coding: [
+        {
+          code: '5090AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
+          display: 'Height (cm)',
+        },
+      ],
+      text: 'Height (cm)',
+    },
+    valueQuantity: {
+      value: 12.0,
+      unit: 'cm',
+    },
+    extension: [
+      {
+        url: 'http://fhir.bahmni.org/ext/observation/form-namespace-path',
+        valueString: 'Bahmni^Vitals (6 years or older).1/10-0',
+      },
+    ],
+  },
+];
+
+export const mockEncounterWithProvider: Encounter = {
+  resourceType: 'Encounter',
+  id: 'encounter-1',
+  status: 'unknown',
+  class: {
+    system: 'http://terminology.hl7.org/CodeSystem/v3-ActCode',
+    code: 'AMB',
+  },
+  subject: {
+    reference: 'Patient/patient-uuid',
+  },
+  participant: [
+    {
+      individual: {
+        reference: 'Practitioner/60b31d2a-1d0c-11f1-b099-5a3ed7acdb7e',
+        type: 'Practitioner',
+        display: 'Super Man',
+      },
+    },
+  ],
+  period: {
+    start: '2026-07-20T09:59:41+00:00',
+  },
+};
+
+export const mockEncounterWithoutProvider: Encounter = {
+  resourceType: 'Encounter',
+  id: 'encounter-2',
+  status: 'unknown',
+  class: {
+    system: 'http://terminology.hl7.org/CodeSystem/v3-ActCode',
+    code: 'AMB',
+  },
+  subject: {
+    reference: 'Patient/patient-uuid',
+  },
+  participant: [],
+  period: {
+    start: '2026-07-20T10:00:00+00:00',
+  },
+};
+
+export const mockObservationAndEncounterBundle: Bundle<
+  Observation | Encounter
+> = {
+  resourceType: 'Bundle',
+  type: 'searchset',
+  total: 3,
+  entry: [
+    {
+      fullUrl: 'http://localhost/openmrs/ws/fhir2/R4/Observation/obs-1',
+      resource: mockObservationWithFormPath,
+    },
+    {
+      fullUrl: 'http://localhost/openmrs/ws/fhir2/R4/Observation/obs-vitals-2',
+      resource: mockObservationsForVitals[1],
+    },
+    {
+      fullUrl: 'http://localhost/openmrs/ws/fhir2/R4/Encounter/encounter-1',
+      resource: mockEncounterWithProvider,
+    },
+  ],
+};
+
+export const mockEmptyObservationsBundle: Bundle<Observation> = {
+  resourceType: 'Bundle',
+  type: 'searchset',
+  total: 0,
+  entry: [],
+};
